Fix CI build errors
(apparently AppVeyor and Travis don't know C# 6 yet...)
Preserve stack trace when rethrowing exception in Cast(Task)
